requirements.txt
torch
python-chess
huggingface_hub
datasets
tqdm
matplotlib
